% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate}
\alias{climate}
\title{Climate data.}
\format{
A list with 40 elements, each element containing the climate data of the respective station:
\itemize{
\item{name}
\item{code}
\item{normTemperature}
\item{latitude}
\item{longitude}
\item{altitude}
\item{climateCode}
\item{situationCode}
\item{monthlyTemperatureAvg}
\item{monthlyIrradiationSum}
\item{PrecipitationSum}
\item{mischungsverhaeltnis}
\item{absFeuchtigkeit}
}
}
\source{
SIA Klimadaten deutsch (kompakt) - Ausgabe 2015, \url{https://www.energytools.ch/index.php/de/downloads/datenbanken}
}
\usage{
climate
}
\description{
A dataset containing climate data for 40 climate stations in Switzerland.
}
\keyword{datasets}
