% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constants}
\alias{constants}
\title{Constants and coefficients.}
\format{
A list with 10 elements, each element containing the parameters of the respective constant/coefficient:
\itemize{
  \item{buildingGenerator}
  \item{oilHeating}
  \item{gasHeating}
  \item{other}
  \item{dhw_systems}
  \item{heating_distribution_systems}
  \item{insulation_material}
  \item{ventilation_recovery}
  \item{insulation_modern}
  \item{insulation_historical}
}
}
\source{
BS2, \url{https://www.bs2.ch/}

Wüest Partner, \url{https://www.wuestpartner.com/}
}
\usage{
constants
}
\description{
Various constants and coefficients required to calculate the CO2 emissions.
}
\keyword{datasets}
