% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Gather and Validate Parameters for CO2 Emission Calculations}
\usage{
get_parameters(
  area,
  floors,
  year,
  utilisation_key,
  climate_code,
  energy_carrier,
  walls_refurb_year = NULL,
  roof_refurb_year = NULL,
  windows_refurb_year = NULL,
  floor_refurb_year = NULL,
  heating_install_year = NULL
)
}
\arguments{
\item{params}{See \code{\link{calculate_emissions}} for details on the parameters.}
}
\value{
A list containing all validated and formatted input parameters, including any
refurbishment details.
}
\description{
This function validates the input parameters and prepare the \code{input_args} list for further processing .
by the \code{\link{getParams}} function. It ensures that all mandatory parameters are present and
correctly formatted, and it collects information about any refurbishments that have been carried out.
}
