% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Q_DHW_Demand}
\alias{Q_DHW_Demand}
\title{Calculate Monthly DHW Energy Demand per Square Meter}
\usage{
Q_DHW_Demand(params)
}
\arguments{
\item{params}{A list returned by the \code{\link{get_parameters}} function.}
}
\value{
A numeric vector of length 12, representing the monthly energy demand
for DHW per square meter, measured in megajoules (MJ). Each element of the
vector corresponds to one month of the year.
}
\description{
\code{Q_DHW_Demand} estimates the monthly energy required for Domestic Hot Water (DHW)
per square meter of the building. The function calculates this demand based on
the daily DHW demand specific to the building type.
}
