% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Q_Gains_Total}
\alias{Q_Gains_Total}
\title{Calculate Total Monthly Heat Gains}
\usage{
Q_Gains_Total(params)
}
\arguments{
\item{params}{A list returned by the \code{\link{get_parameters}} function.}
}
\value{
A numeric vector of length 12, representing the total heat gains for each month,
  measured in megajoules per square meter (MJ/m²).
}
\description{
`Q_Gains_Total` aggregates the total heat gains within a building per square meter, 
combining electrical gains, body heat gains, and solar gains.
}
