% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{simpleWindowElements}
\alias{simpleWindowElements}
\title{Generate Simple Window Elements for a Building}
\usage{
simpleWindowElements(params)
}
\arguments{
\item{params}{A list containing parameters for the building, including:
\describe{
  \item{uValuesBefore}{A list of U-values for various building elements before any refurbishment.}
  \item{area}{Total floor area of the building.}
  \item{floors}{Number of floors in the building.}
}}
}
\value{
A list of window elements where each element is a list containing:
  \describe{
    \item{name}{Name of the window element, indicating its type and orientation.}
    \item{orientation}{Orientation of the window ('H' for roof windows, 'S', 'W', 'N', 'E' for side windows).}
    \item{uValue}{The thermal transmittance (U-value) of the window.}
    \item{area}{The area of the window in square meters.}
    \item{opacity}{Light opacity of the window. Constant at 60\% for all windows in this simplified model.}
    \item{frameFactor}{The frame factor, representing the proportion of the window area that is framing. Constant at 70\% for all windows in this simplified model.}
    \item{shadingFactor}{A factor representing the effect of shading devices on the window. Constant at 80\% for all windows in this simplified model.}
  }
}
\description{
`simpleWindowElements` calculates and generates a list of window elements for a building based on provided
parameters and predefined constants. This includes roof windows and side windows with specific characteristics
such as orientation, thermal transmittance (U-values), area, and factors affecting light and heat transmission
through the windows. This function is useful for estimating the thermal and lighting characteristics of a
building's window elements in a simplified manner.
}
\details{
The function assumes standard building geometries, applies fixed ratios for window-to-wall area
and roof angles and standard factors for windows properties. It is designed for preliminary assessments and may not substitute detailed thermal or
lighting modeling.
}
