% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utilisation}
\alias{utilisation}
\title{Utilisation.}
\format{
A list with 12 elements, each element containing the parameters of the respective utilisation:
\itemize{
\item{key}
\item{name}
\item{temperature}
\item{personArea}
\item{personHeat}
\item{personPresence}
\item{electricityUse}
\item{electricityUseFactor}
\item{ventilationRate}
\item{dhwDemand}
\item{temperatureAddition}
\item{heatgainsFactor}
\item{heatgainsTime}
}
}
\source{
Thermische Energie im Hochbau, SIA 380/1:2009, \url{http://www.sia.ch/}
}
\usage{
utilisation
}
\description{
Building utilisations according to SIA 380/1:2009.
}
\keyword{datasets}
