% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Q_Heating_Demand}
\alias{Q_Heating_Demand}
\title{Calculate Monthly Heating Energy Demand}
\usage{
Q_Heating_Demand(params)
}
\arguments{
\item{params}{A list returned by the \code{\link{get_parameters}} function.}
}
\value{
A numeric vector of length 12, representing the net heating energy demand
for each month, measured in megajoules (MJ) per square meter. Each element of
the vector corresponds to one month of the year, providing a detailed view of
the heating demand's seasonal variation.
}
\description{
The function applies formula 117 of the SIA 380/1 norm, and aligns with varaible \code{Q_h}.\cr
\code{Q_Heating_Demand} computes the net energy demand for heating on a monthly basis,
per square meter of the building.
This function calculates the heating demand by assessing the
difference between total heat losses and gains. It utilizes \code{\link{Q_Losses_Total}}
for total heat losses and \code{\link{Q_Gains_Used}} for usable heat gains.
}
