% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Q_Losses_Transmission}
\alias{Q_Losses_Transmission}
\title{Calculate Monthly Transmission Heat Losses in MJ/m2}
\usage{
Q_Losses_Transmission(params)
}
\arguments{
\item{params}{A list returned by the \code{\link{get_parameters}} function.}
}
\value{
A numeric vector of length 12, representing the total transmission heat losses for each month,
expressed in MJ/m2. Negative loss values are set to 0 to ensure that all output values represent actual heat losses.
}
\description{
The function applies formulas 76-96 of the SIA 380/1 norm, and aligns with varaible \code{Q_T}. \cr
\code{Q_Losses_Transmission} calculates the heat transmission losses through the building envelope for each month,
converting the results into megajoules per square meter (MJ/m2). It assess losses based on varying outside temperatures throughout the year. It utilizes the
\code{\link{calculateTransmissionLosses}} function to compute losses for each building element (opaque and window elements),
then aggregates these losses to provide a monthly total.
}
